% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgestan_helpers.R
\name{compile_bs}
\alias{compile_bs}
\title{Write R data to bridgestan data JSON file}
\usage{
compile_bs(model_name, bridgestan_path)
}
\arguments{
\item{model_name}{Name of the bridgestan folder containing the .stan file to be compiled}

\item{bridge_stan_path}{Path to Bridgestan dir}
}
\description{
Compiling a stan model into a bridgestan binary shared object (.so) file
}
