% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgestan_helpers.R
\name{get_mixture_data}
\alias{get_mixture_data}
\title{Generation of mixture normal data}
\usage{
get_mixture_data(n_samples, pi, mu, sigma)
}
\arguments{
\item{n_samples}{Number of samples}

\item{pi}{Vector of class probabilities}

\item{mu}{Maxtrix of cluster means}

\item{sigma}{Matrix of cluster standard deviationos}
}
\description{
Generates a data set distributed according to a normal mixture distribution
}
