% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{get_trajectory_stats}
\alias{get_trajectory_stats}
\title{Sample trajectory summary generator}
\usage{
get_trajectory_stats(sim, n_trajectories = 1)
}
\arguments{
\item{sim}{Input NUTS simulation}

\item{n_trajectories}{Number of trajectories which should be extracted}
}
\description{
Extracting and summarizing parameter, momentum, gradient and hessian
trajectories for a given NUTS-simulation
}
