% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUTS_helpers.R
\name{initialize_model}
\alias{initialize_model}
\title{Initialization of HMC model parameters}
\usage{
initialize_model(model, warm_up)
}
\arguments{
\item{model}{Underlying cmdstan model object}

\item{warmup}{Burn-in-period of the initialization}
}
\description{
Using Stan to initialize the mass matrix, step size and initial parameter value.
}
