% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUTS_helpers.R
\name{leapfrog_1step}
\alias{leapfrog_1step}
\title{One-step integration with leapfrog algorithm}
\usage{
leapfrog_1step(model, current_q, current_p, eps, mass = 1)
}
\arguments{
\item{model}{Underlying bridgestan model object}

\item{current_q}{Vector with current parameter variable values}

\item{eps}{Step size}

\item{mass}{Particle mass for parameter update step}

\item{model_name}{Vector with current momentum variable values}
}
\description{
Conducting one step of the leapfrog integrator. Return updated variables together with the gradient
corresponding to the second momentum update.
}
