% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUTS.R
\name{NUTS}
\alias{NUTS}
\title{NUTS/bridgestan sampler}
\usage{
NUTS(model, n_samples, eps, theta_0, warm_up = floor(n_samples/2))
}
\arguments{
\item{model}{Input bridgestan model}

\item{n_samples}{Number of post warmup samples generated}

\item{eps}{Step size}

\item{theta_0}{Initial parameter value}
}
\description{
Sampling from an input bridgestan model, using the NUTS sampler
}
