% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgestan_helpers.R
\name{write_json_data}
\alias{write_json_data}
\title{Write R data to bridgestan data JSON file}
\usage{
write_json_data(data, model_name, bridgestan_path = BS_PATH)
}
\arguments{
\item{data}{List of data to be converted}

\item{model_name}{Name of the Bridgestan model corresponding tothe data}

\item{bridge_stan_path}{Path to Bridgestan dir}
}
\description{
Converting a list of data into a JSON file recognized by the bridgestan compiler
}
