% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUTS.R
\name{NUTS}
\alias{NUTS}
\title{NUTS/bridgestan sampler}
\usage{
NUTS(
  model_object,
  n_samples,
  warm_up = floor(n_samples/2),
  initialize_model = TRUE,
  return_hessian = TRUE,
  initial_params = NULL
)
}
\arguments{
\item{n_samples}{Number of post warmup samples generated}

\item{warm_up}{Sampler burn-in-period}

\item{initialize_model}{TRUE if stan's initialization of mass matrix and step size should be used}

\item{return_hessian}{TRUE if the full hessian should be genererated at each leapfrog iteration}

\item{initial_params}{Optional list containing initial custom parameters "step_size", "mass_matrix" and "theta_0".
Used when initialize_model == FALSE}

\item{model}{Input bridgestan and cmdstan model object}
}
\description{
Sampling from an input bridgestan model, using the NUTS sampler
}
