% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUTS.R
\name{NUTS}
\alias{NUTS}
\title{NUTS/bridgestan sampler}
\usage{
NUTS(stan_model, theta_0, eps, n_samples)
}
\arguments{
\item{stan_model}{Input bridgestan model}

\item{theta_0}{Initial parameter value}

\item{eps}{Step size}

\item{n_samples}{Number of post warmup samples generated}
}
\description{
Sampling from an input bridgestan model, using the NUTS sampler
}
